recommand_place/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
gensim>=4.1.0
python-dotenv>=0.19.0
openai>=1.0.0
python-multipart>=0.0.5
geopy>=2.2.0 